% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-functions.R
\name{filterCounts}
\alias{filterCounts}
\title{Filter matrix for low counts}
\usage{
filterCounts(
  object,
  data,
  y,
  min.count = 10,
  min.total.count = 15,
  large.n = 10,
  min.prop = 0.7
)
}
\arguments{
\item{object}{an object of an appropriate class.  For the default
    method, a model \link[stats]{formula} or a \code{\link[stats]{terms}} object.}

\item{data}{a data frame created with \code{\link[stats]{model.frame}}.  If
    another sort of object, \code{model.frame} is called first.}

\item{y}{matrix of counts, or a \code{DGEList} object, or a \code{SummarizedExperiment} object.}

\item{min.count}{numeric. Minimum count required for at least some samples.}

\item{min.total.count}{numeric. Minimum total count required.}

\item{large.n}{integer. Number of samples per group that is considered to be \dQuote{large}.}

\item{min.prop}{numeric. In large sample situations, the minimum proportion of samples in a group that a gene needs to be expressed in. See Details below for the exact formula.}
}
\value{
filter and transposed count matrix
}
\description{
Wrapper for model.matrix and filterByExpr for removing low counts genes from expression matrix
}
