% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-functions.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalize and log-transform count matrix}
\usage{
normalizeCounts(counts, log = TRUE, prior.count = 5, ...)
}
\arguments{
\item{counts}{numeric matrix or data.frame containing sequence read counts, with rows corresponding to genes (genomic features) and columns to samples. Negative values or NAs are not allowed.}

\item{log}{logical, if \code{TRUE} then \code{log2} values are returned.}

\item{prior.count}{average count to be added to each observation to avoid taking log of zero. Used only if \code{log=TRUE}.}

\item{...}{additional arguments passed to edgeR::calcNormFactors}
}
\value{
transposed and log-normalized expression matrix
}
\description{
Wrapper for calcNormFactors and cpm from the edgeR package.
}
