% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plotTotal}
\alias{plotTotal}
\alias{plotUnconstrained}
\alias{plotConstrained}
\alias{plotRDA}
\alias{plotDecomposition}
\title{Visualize RDA or CCA with variance partitioning}
\usage{
plotTotal(object, color = NULL, shape = NULL, size = 3)

plotUnconstrained(object, color = NULL, shape = NULL, size = 3)

plotConstrained(object, color = NULL, shape = NULL, size = 3)

plotRDA(object, color = NULL, shape = NULL, size = 3)

plotDecomposition(
  object,
  color = NULL,
  shape = NULL,
  size = 3,
  color_scale = NULL,
  shape_scale = NULL
)
}
\arguments{
\item{object}{rda or cca object}

\item{color}{vector}

\item{shape}{vector}

\item{size}{scalar}

\item{color_scale}{Scale object from ggplot2}

\item{shape_scale}{Scale object from ggplot2}
}
\value{
ggplot
}
\description{
Visualize RDA or CCA with variance partitioning
}
\section{Functions}{
\itemize{
\item \code{plotTotal()}: Plot total variance

\item \code{plotUnconstrained()}: Plot unconstrained variance

\item \code{plotConstrained()}: Plot constrained variance

\item \code{plotRDA()}: Plot RDA

}}
