% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plotDesign}
\alias{plotDesign}
\title{Visualize RDA or CCA with study design}
\usage{
plotDesign(
  object,
  color = NULL,
  shape = NULL,
  size = 3,
  bp_col = "black",
  bp_scale = 0.9
)
}
\arguments{
\item{object}{rda or cca object}

\item{color}{vector}

\item{shape}{vector}

\item{size}{scalar}

\item{bp_col}{character: Color or biplot colors}

\item{bp_scale}{scalar: scaling factor of biplot scale}
}
\description{
Visualize RDA or CCA with study design
}
