% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plotStrain}
\alias{plotStrain}
\title{Visualize RDA or CCA design strain (spider plot)}
\usage{
plotStrain(object, color = NULL, shape = NULL, size = 3)
}
\arguments{
\item{object}{rda or cca object}

\item{color}{vector}

\item{shape}{vector}

\item{size}{scalar}
}
\description{
Visualize RDA or CCA design strain (spider plot)
}
