% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-functions.R
\name{stabilizeCounts}
\alias{stabilizeCounts}
\title{Variance-stabilize count matrix}
\usage{
stabilizeCounts(design, colData, countData, blind = FALSE)
}
\arguments{
\item{design}{a \code{formula} or \code{matrix}.
the \code{formula} expresses how the counts for each gene
depend on the variables in \code{colData}. Many R \code{formula} are valid,
including designs with multiple variables, e.g., \code{~ group + condition},
and designs with interactions, e.g., \code{~ genotype + treatment + genotype:treatment}.
See \code{\link[DESeq2]{results}} for a variety of designs and how to extract results tables.
By default, the functions in this package will use 
the last variable in the formula for building results tables and plotting.
\code{~ 1} can be used for no design, although users need to remember
to switch to another design for differential testing.}

\item{colData}{for matrix input: a \code{DataFrame} or \code{data.frame} with at least a single column.
Rows of colData correspond to columns of countData}

\item{countData}{for matrix input: a matrix of non-negative integers}

\item{blind}{logical, whether to blind the transformation to the experimental
design. blind=TRUE should be used for comparing samples in a manner unbiased by
prior information on samples, for example to perform sample QA (quality assurance).
blind=FALSE should be used for transforming data for downstream analysis,
where the full use of the design information should be made.
blind=FALSE will skip re-estimation of the dispersion trend, if this has already been calculated.
If many of genes have large differences in counts due to
the experimental design, it is important to set blind=FALSE for downstream
analysis.}
}
\value{
transposed and log-normalized expression matrix
}
\description{
Wrapper for varianceStabilizingTransformation from the DESeq2 package.
}
