% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veganomics-package.R
\docType{package}
\name{veganomics-package}
\alias{veganomics}
\alias{veganomics-package}
\title{veganomics: vegan for -omics data}
\description{
Plotting and utility functions for using vegan with -omics data.
}
\keyword{internal}
